+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.0.2
 
+- Feature: skip entity class creation on demand, only serialization generation (attribute generateEntity="false")
 - Feature: custom imports defined on xml level (import node)
 - Feature: type now can be given in xml node name, no need to use field node
 - Feature: add comment to generated field (attribute comment)